Move plugin.Registry to a separate file
plugin/types.go is meant to contain things that plugin authors can
use when writing plugins. plugin.Registry is not something plugin
authors should know about, so this commit moves it to its own file.
